+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@myobie/pty",
-  "version": "0.11.0",
+  "version": "0.12.0",
   "description": "Persistent terminal sessions with detach/attach, plus a Playwright-style testing library for TUI apps",
   "type": "module",
   "license": "MIT",
